requirements.txt
flask==2.3.3
openai==1.3.7
anthropic==0.8.1
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
